requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==21.2.0
chardet==3.0.4
discord.py==1.5.0
idna==3.2
multidict==4.7.6
typing-extensions==3.10.0.0
yarl==1.5.1
